requirements.txt
streamlit>=1.24.0
pdfminer.six>=20201018
python-docx>=0.8.11
spacy>=3.5.0
reportlab>=3.6.12
rapidfuzz>=2.13.7
pandas>=1.5.0
Pillow>=9.0.0
